[Wallet]: Move Single Use Components to Feature Directories
