+++ components/brave_wallet_ui/components/desktop/popup-modals/edit-visible-assets-modal/components/visible_asset_item/visible_asset_item.style.ts
@@ -1,7 +1,8 @@
 // Copyright (c) 2021 The Brave Authors. All rights reserved.
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
 import styled from 'styled-components'
 import * as leo from '@brave/leo/tokens/css/variables'
 import LeoIcon from '@brave/leo/react/icon'
@@ -11,10 +12,12 @@ import {
   AssetIconProps,
   AssetIconFactory,
   WalletButton,
-} from '../../shared/style'
-import { layoutPanelWidth } from '../wallet-page-wrapper/wallet-page-wrapper.style'
+} from '$wallet/components/shared/style'
+import {
+  layoutPanelWidth, //
+} from '$wallet/components/desktop/wallet-page-wrapper/wallet-page-wrapper.style'
 
-export const assetWatchListItemHeight = 58
+export const visibleAssetItemHeight = 58
 
 export const StyledWrapper = styled.div`
   display: flex;
