+++ components/brave_wallet_ui/market/components/assets_filter_dropdown/assets_filter_dropdown.style.ts
@@ -1,13 +1,13 @@
 // Copyright (c) 2022 The Brave Authors. All rights reserved.
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
+// You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import styled from 'styled-components'
 import * as leo from '@brave/leo/tokens/css/variables'
 
-import CaratDownIcon from '../../../assets/svg-icons/carat-down.svg'
-import { WalletButton } from '../../shared/style'
+import CaratDownIcon from '$wallet/assets/svg-icons/carat-down.svg'
+import { WalletButton } from '$wallet/components/shared/style'
 
 export const StyledWrapper = styled.div`
   display: flex;
