+++ components/brave_wallet_ui/market/components/assets_filter_dropdown/components/assets_filter_option/assets_filter_option.tsx
@@ -1,11 +1,11 @@
 // Copyright (c) 2022 The Brave Authors. All rights reserved.
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
+// You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import * as React from 'react'
 
-import { Option } from './style'
+import { Option } from './assets_filter_option.style'
 
 export interface Props {
   label: string
