+++ components/brave_wallet_ui/page/screens/backup-wallet/backup-recovery-phrase/backup-recovery-phrase.tsx
@@ -27,7 +27,7 @@ import { useSafePageSelector } from '../../../../common/hooks/use-safe-selector'
 // components
 import {
   RecoveryPhrase, //
-} from '../../../../components/desktop/recovery-phrase/recovery-phrase'
+} from './components/recovery_phrase/recovery_phrase'
 import {
   OnboardingContentLayout, //
 } from '../../onboarding/components/onboarding_content_layout/content_layout'
