+++ components/brave_wallet_ui/page/screens/backup-wallet/backup-recovery-phrase/components/recovery_phrase/recovery_phrase.style.tsx
@@ -1,14 +1,14 @@
 // Copyright (c) 2022 The Brave Authors. All rights reserved.
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
+// You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import styled from 'styled-components'
 import Icon from '@brave/leo/react/icon'
 import * as leo from '@brave/leo/tokens/css/variables'
 import {
   layoutSmallWidth, //
-} from '../wallet-page-wrapper/wallet-page-wrapper.style'
+} from '$wallet/components/desktop/wallet-page-wrapper/wallet-page-wrapper.style'
 
 export const RecoveryPhraseContainer = styled.div<{ phraseLength: number }>`
   display: grid;
