+++ components/brave_wallet_ui/page/screens/backup-wallet/backup-recovery-phrase/components/recovery_phrase/recovery_phrase.tsx
@@ -1,12 +1,12 @@
 // Copyright (c) 2022 The Brave Authors. All rights reserved.
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
+// You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import * as React from 'react'
 
 // utils
-import { unbiasedRandom } from '../../../utils/random-utils'
+import { unbiasedRandom } from '$wallet/utils/random-utils'
 
 // styles
 import {
@@ -16,7 +16,7 @@ import {
   HiddenPhraseContainer,
   EyeOffIcon,
   WordPos,
-} from './recovery-phrase.style'
+} from './recovery_phrase.style'
 
 export interface SelectedPhraseWord {
   index: number
