+++ components/brave_wallet_ui/page/screens/onboarding/components/steps_navigation/steps_navigation.stories.tsx
@@ -1,12 +1,12 @@
 // Copyright (c) 2022 The Brave Authors. All rights reserved.
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
-// you can obtain one at https://mozilla.org/MPL/2.0/.
+// You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import * as React from 'react'
 import { MemoryRouter as Router, Route } from 'react-router'
 
-import { StepsNavigation } from './steps-navigation'
+import { StepsNavigation } from './steps_navigation'
 
 export const _StepsNavigation = {
   render: () => (
